requirements.txt
flask
flask-cors
cloudinary
python-dotenv
basicsr==1.4.2
realesrgan
torch==2.0.1+cpu
torchvision==0.15.2+cpu
opencv-python-headless
Pillow
numpy<2